correcting some typoes
the project compiles

+++ calculator.h
@@ -1,10 +1,11 @@
 #ifndef _CALCULATOR_H_
 #define _CALCULATOR_H_
 
+#include <stdint.h>
 #include <avr/io.h>
-##include <util/delay.h>
+#include <util/delay.h>
 #include "LCD/LCD.h"
-#include "Keypad/Keypad.h"
+#include "keypad/keypad.h"
 #include "Stack/stack.h"
 
 
@@ -13,7 +14,7 @@
 uint8_t* read_exp(void);
 uint16_t string_to_number(uint8_t*);
 uint8_t* number_to_string(uint16_t);
-uint8_t infix_exp_eval(uint8_t*);
+uint16_t infix_exp_eval(uint8_t*);
 uint16_t clac(uint8_t,uint8_t);
 void print_result(uint16_t);
